% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinci.R
\name{within_ci}
\alias{within_ci}
\title{Within Subjects Confidence Intervals.}
\usage{
within_ci(data, subject, value, conf = 0.95, ws_factors = c(),
  bs_factors = c())
}
\arguments{
\item{data:}{Dataframe in long format}

\item{subject:}{Name of single column in data containing participant identifiers}

\item{value:}{Name of single column in data containing values of DV}

\item{conf:}{Confidence interval to be calculated}

\item{ws_factors:}{names of (multiple) columns in data containing within-subject factors}

\item{bs_factors:}{names of (multiple) columns in data containing between-subject factors}
}
\value{
Returns length of Confidence Intervals per factor
}
\description{
\code{within_ci} returns lower and upper bounds of confidence intervals
per factor for data with at least one within subject factor.
}
\examples{
# One within factor, no between factors
within_ci(data = data_1w,
           subject = subject,
           value = DV,
           ws_factors = within1)

# Multiple within and between factors
within_ci(data = data_2w2b,
           subject = subject,
           value = DV,
           ws_factors = c(within1, within2),
           bs_factors = c(between1, between2))

# Multiple within and between factors with graph
 data_2w2b \%>\%
   within_ci(data = .,
             subject = subject,
             value = DV,
             ws_factors = c(within1, within2),
             bs_factors = c(between1, between2)) \%>\%
  ggplot(aes(x = within1,
             y = sample_mean,
             group = within2)) +
  geom_bar(stat = "identity",
           position = position_dodge(width = 1),
           color = "black",
           aes(fill = within2)) +
  geom_errorbar(position = position_dodge(width = 1),
                width = .1,
                aes(ymin = sample_mean - CI,
                    ymax = sample_mean + CI)) +
  scale_fill_manual(values = c("white", "grey")) +
  labs(y = "DV",
       title = "Example bar plot - two within two between") +
  theme_classic() +
  facet_grid(between1 ~ between2)

}
